feat: add POST /api/customers endpoint (#37)
The purpose of this change is to allow us to introduce a delay in the response times on demand. The new `POST /api/customers` endpoint expects a body containing a new customer (in the form of a JSON document).

Here's an example JSON document for creating a new customer:

```json
{
  "full_name": "Catherine Weaver",
  "company_name": "Cyberdyne Systems",
  "email": "cw@example.com",
  "address": "18144 El Camino Real",
  "postal_code": "5421",
  "city": "Sunnyvale",
  "country": "USA"
}
```

The code will validate that the e-mail is a valid e-mail using a regular expression. This regex however is vulnerable to a so called [ReDos](https://www.owasp.org/index.php/Regular_expression_Denial_of_Service_-_ReDoS) attack, so if you instead of a normal e-mail like `foo@example.com`, send up the string `aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa!`, the regular expression will go into a very expensive loop that will make the server freeze up for a few seconds. The more `a`'s you add the the string the longer the server will freeze up. While the server is spending time trying to validate the e-mail, it can't respond to any other requests, so the overall server response time will shoot up.

This commit also adds two npm scripts:

- `customer-add-ok`, which will add a new customer
- `customer-add-redos`, which fail adding a new customer and slow down
  the server in the process

Run either of the two scripts using `npm run <name>`.