chore: make it the responsibility of opbeans-node to update its APM agent dep (#163)
This adds a second "Opbeans Node Bump" Jenkins job that runs weekly on the
"main" branch. It checks for an available agent update (a published version
newer than what is in the current package-lock file), and if there is one it:
bumps to that ver, pushes, and git tags with "v$VERSION". The push and tag
will trigger the usual opbeans Jenkins pipeline to publish docker images.

Refs: https://github.com/elastic/apm-agent-nodejs/issues/2728
Fixes: #164
Co-authored-by: Victor Martinez <VictorMartinezRubio@gmail.com>